// packages/credentials/src/V1/types.ts
export interface Proof {
  type: string
}

export interface UnsignedVc {
  /**
   * References to json-ld contexts defining the terms used.
   */
  '@context': [typeof W3C_CREDENTIAL_CONTEXT_URL, ...string[]]
  /**
   * Credential identifier.
   */
  id?: string
  /**
   * The credential types, which declare what data to expect in the credential.
   */
  type: typeof W3C_CREDENTIAL_TYPE | string[]
  /**
   * Claims about the subjects of the credential.
   */
  credentialSubject: {
    id?: string
    [key: string]: unknown
  }
  /**
   * The entity that issued the credential.
   */
  issuer: string
  /**
   * When the credential was issued.
   */
  issuanceDate?: string
  /**
   * If true, this credential can only be presented and used by its subject.
   */
  nonTransferable?: boolean
  /**
   * Contains json schema for the validation of credentialSubject claims.
   */
  credentialSchema?: {
    id?: string
    type: string
  }
  /**
   * Contains credentials status method.
   */
  credentialStatus?: {
    id: string
    type: string
  }
}

export interface VerifiableCredential extends UnsignedVc {
  issuanceDate: string
  /**
   *  Cryptographic proof that makes the credential tamper-evident.
   */
  proof: Proof | Proof[]
}

export interface VerifiablePresentation {
  id?: string
  '@context': [typeof W3C_CREDENTIAL_CONTEXT_URL, ...string[]]
  type: [typeof W3C_PRESENTATION_TYPE, ...string[]]
  verifiableCredential: VerifiableCredential | VerifiableCredential[]
  holder: Did
  proof?: Proof | Proof[]
  expirationDate?: string
  issuanceDate?: string
  verifier?: string
}